requirements.txt
streamlit==1.47.1
google-genai>=1.32.0
python-dotenv==1.0.0